+++ src/main/java/com/grafana/opentelemetry/ConnectionProperties.java
@@ -1,22 +1,7 @@
 package com.grafana.opentelemetry;
 
+import java.time.Duration;
 import java.util.Map;
 import java.util.Optional;
 
-class ConnectionProperties {
-  private final Map<String, String> headers;
-  private final Optional<String> endpoint;
-
-  public ConnectionProperties(Optional<String> endpoint, Map<String, String> headers) {
-    this.headers = headers;
-    this.endpoint = endpoint;
-  }
-
-  public Map<String, String> getHeaders() {
-    return headers;
-  }
-
-  public Optional<String> getEndpoint() {
-    return endpoint;
-  }
-}
+public record ConnectionProperties(Optional<String> endpoint, Map<String, String> headers, Duration metricExportInterval) {}