+++ src/main/java/com/grafana/opentelemetry/MetricsOtlpConfig.java
@@ -1,6 +1,9 @@
 package com.grafana.opentelemetry;
 
+import io.micrometer.core.instrument.config.validate.DurationValidator;
 import io.micrometer.registry.otlp.OtlpConfig;
+
+import java.time.Duration;
 import java.util.Map;
 import java.util.concurrent.TimeUnit;
 
@@ -23,14 +26,14 @@ class MetricsOtlpConfig implements OtlpConfig {
   @Override
   public String url() {
     return connectionProperties
-        .getEndpoint()
+        .endpoint()
         .map(s -> s + "/v1/metrics")
         .orElse(OtlpConfig.DEFAULT.url());
   }
 
   @Override
   public Map<String, String> headers() {
-    return connectionProperties.getHeaders();
+    return connectionProperties.headers();
   }
 
   @Override
@@ -41,5 +44,10 @@ class MetricsOtlpConfig implements OtlpConfig {
   @Override
   public TimeUnit baseTimeUnit() {
     return TimeUnit.SECONDS;
+  }
+
+  @Override
+  public Duration step() {
+    return connectionProperties.metricExportInterval();
   }
 }